Requested changes Michelle
This call aims to identify, recognise, and reward passionate members of the R community, including people who belong to groups that are historically and systematically excluded. This is a 12-month program.

And please add another row at the end of the past funding  (as now it's a bit short 🙂) saying: More past calls coming soon.